hotshot_libp2p_networking::network::behaviours
Module dht

Wrapper around Kademlia

Modules

bootstrap

Task for doing bootstraps at a regular interval

record

Additional DHT record functionality

store

Additional DHT store functionality

Structs

Bootstrap

State of bootstrapping

DHTBehaviour

Behaviour wrapping libp2p’s kademlia included:

KadGetQuery 🔒

Metadata holder for get query

KadPutQuery

Metadata holder for get query

MAX_DHT_QUERY_SIZE 🔒

the maximum number of nodes to query in the DHT at any one time

Enums

DHTEvent

DHT event enum

DHTProgress

represents progress through DHT

State

State used for random walk and bootstrapping

Constants

NUM_REPLICATED_TO_TRUST 🔒

the number of nodes required to get an answer from in order to trust that the answer is correct when retrieving from the DHT