hotshot_libp2p_networking::network::behaviours::dht
Module record

Additional DHT record functionality

Structs

RecordKey

A record’s key. This is a concatenation of the namespace and the key.

Enums

Namespace

The namespace of a record. This is included with the key and allows for multiple types of records to be stored in the DHT.

RecordValue

A (signed or unsigned) record value to be stored (serialized) in the DHT. This is a wrapper around a value that includes a possible signature.

Functions

requires_authentication 🔒

Require certain namespaces to be authenticated