hotshot_libp2p_networking::network::node
Module handle

libp2p network handle allows for control over the libp2p network

Structs

NetworkNodeHandle

A handle containing:

NetworkNodeReceiver

internal network node receiver

Functions

spawn_network_node

Spawn a network node task task and return the handle and the receiver for it