hotshot_libp2p_networking::network
Module transport

Alternative Libp2p transport implementations

Structs

AuthMessage 🔒

The deserialized form of an authentication message that is sent to the remote peer

StakeTableAuthentication

A wrapper for a Transport that bidirectionally authenticates connections by performing a handshake that checks if the remote peer is present in the stake table.

Constants

AUTH_HANDSHAKE_TIMEOUT 🔒

The timeout for the authentication handshake. This is used to prevent attacks that keep connections open indefinitely by half-finishing the handshake.

MAX_AUTH_MESSAGE_SIZE 🔒

The maximum size of an authentication message. This is used to prevent DoS attacks by sending large messages.

Traits

AsOutput 🔒

A helper trait that allows us to access the underlying connection and PeerId from a transport output

Functions

construct_auth_message

Create an sign an authentication message to be sent to the remote peer

read_length_delimited

A helper function to read a length-delimited message from a stream. Takes into account the maximum message size.

write_length_delimited

A helper function to write a length-delimited message to a stream.

Type Aliases

UpgradeFuture 🔒

A type alias for the future that upgrades a connection to perform the authentication handshake