hotshot_libp2p_networking::network
Module node

functionality of a libp2p network node

Re-exports

pub use self::config::GossipConfig;

pub use self::config::NetworkNodeConfig;

pub use self::config::NetworkNodeConfigBuilder;

pub use self::config::NetworkNodeConfigBuilderError;

pub use self::config::RequestResponseConfig;

pub use self::config::DEFAULT_REPLICATION_FACTOR;

pub use self::handle::spawn_network_node;

pub use self::handle::NetworkNodeHandle;

pub use self::handle::NetworkNodeReceiver;

Modules

config 🔒

configuration for the libp2p network (e.g. how it should be built)

handle 🔒

libp2p network handle allows for control over the libp2p network

Structs

NetworkNode

Network definition

Constants

ESTABLISHED_LIMIT

Wrapped num of connections

ESTABLISHED_LIMIT_UNWR

Number of connections to a single peer before logging an error

MAX_GOSSIP_MSG_SIZE

Maximum size of a message